+++ BackEnd/src/main/java/com/example/backend/repo/StoreRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StoreRepo extends JpaRepository<Store, Integer> {
+public interface StoreRepo extends JpaRepository<Store, Long> {
 }
